Create Book Service with DTO classes
