--- BusinessLogicLayer/DataTransferObjects/AuthorDTO.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace BusinessLogicLayer
-{
-    public class AuthorDTO
-    {
-        public int Id { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public DateTime? BirthDate { get; set; }
-    }
-}
