+++ DataAccessLayer/Repositories/GenericRepository.cs
@@ -23,7 +23,7 @@ namespace DataAccessLayer
         public virtual IEnumerable<TEntity> Get(
             Expression<Func<TEntity, bool>> filter = null,
             Func<IQueryable<TEntity>, IOrderedQueryable<TEntity>> orderBy = null,
-            string includeProperties = "")
+            string includeProperties = "") // params string[] includeProperties
         {
             IQueryable<TEntity> query = dbSet;
