+++ LibraryApp.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PresentationLayer", "Presen
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataAccessLayer", "DataAccessLayer\DataAccessLayer.csproj", "{2AD695CB-4B84-4592-8B28-E42FA34CBC0F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogicLayer", "BusinessLogicLayer\BusinessLogicLayer.csproj", "{7BA67748-FDA5-4B0B-AC75-0CD194887881}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogicLayer", "BusinessLogicLayer\BusinessLogicLayer.csproj", "{08CBCE6A-AF66-4E62-B73A-CEC9A82D4C11}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{2AD695CB-4B84-4592-8B28-E42FA34CBC0F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2AD695CB-4B84-4592-8B28-E42FA34CBC0F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2AD695CB-4B84-4592-8B28-E42FA34CBC0F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7BA67748-FDA5-4B0B-AC75-0CD194887881}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BA67748-FDA5-4B0B-AC75-0CD194887881}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BA67748-FDA5-4B0B-AC75-0CD194887881}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BA67748-FDA5-4B0B-AC75-0CD194887881}.Release|Any CPU.Build.0 = Release|Any CPU
+		{08CBCE6A-AF66-4E62-B73A-CEC9A82D4C11}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{08CBCE6A-AF66-4E62-B73A-CEC9A82D4C11}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{08CBCE6A-AF66-4E62-B73A-CEC9A82D4C11}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{08CBCE6A-AF66-4E62-B73A-CEC9A82D4C11}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
